+++ CHANGES.rst
@@ -42,6 +42,7 @@ Breaking changes
 * Note by each function call: rename ``data_model`` into ``imodel`` e.g. imodel=icoads_r300_d704 (:pull:`103`)
 * ``cdm_mapper.map_model``: call with (data, imodel=imodel) (:pull:`103`)
 * ``mdf_reader.read``: call with (source, imodel=imodel) (:pull:`103`)
+* ``operations``: delete corrections module (:pull:`104`)
 
 Internal changes
 ^^^^^^^^^^^^^^^^
@@ -60,6 +61,7 @@ Internal changes
 * ``mdf_reader``: replace empty strings with missing values (:pull:`89`)
 * ``metmetpy``: use function ``overwrite_data`` in all platform type correction functions (:pull:`89`)
 * rename ``data_model`` into ``imodel`` (:pull:`103`)
+* implement assertion tests for module operations (:pull:`104`)
 
 Bug fixes
 ^^^^^^^^^